Alguns ajustes

+++ src/main/resources/db/migration/V4__create-table-schedules.sql
@@ -2,7 +2,7 @@ CREATE TABLE tb_schedules (
     id BIGINT NOT NULL PRIMARY KEY,
     date DATE NOT NULL,
     time VARCHAR(50) NOT NULL,
-    price FLOAT NOT NULL,
+    price DECIMAL(6, 2) NOT NULL,
     status ENUM('CONFIRMADO', 'CONCLUIDO', 'CANCELADO') NOT NULL,
     barber_id BIGINT NOT NULL,
     client_id BIGINT NOT NULL,
